% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_search.R
\name{search_wos}
\alias{search_wos}
\title{Run a search query to the Web of Science Web Services Lite API}
\usage{
search_wos(
  sid,
  query = "",
  api = "lite",
  timespan = "",
  editions = if (api == "lite") c("SCI", "ISTP", "ISSHP", "IC"),
  url = "http://search.webofknowledge.com"
)
}
\arguments{
\item{sid}{session id generated by a previous call to \code{\link{wos_authenticate}}}

\item{query}{query text}

\item{api}{the API type. Only "lite" is supported for now}

\item{editions}{search collection editions to query against}

\item{url}{url of WoS service (to be used with a proxy)}
}
\value{
a list with two elements : the number of results, and the generated query id.
}
\description{
Run a search query to the Web of Science Web Services Lite API
}
\seealso{
\code{\link{wos_authenticate}}, \code{\link{wos_retrieve}}
}
